requirements.txt
alembic==1.15.2
Babel==2.17.0
blinker==1.9.0
Brotli==1.1.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cssselect2==0.8.0
exceptiongroup==1.3.0
flask==3.1.1
Flask-Migrate==4.1.0
flask-sqlalchemy==3.1.1
fonttools==4.58.0
greenlet==3.2.2
idna==3.10
importlib-metadata==8.7.0
iniconfig==2.1.0
itsdangerous==2.2.0
jinja2==3.1.6
mako==1.3.10
MarkupSafe==3.0.2
packaging==25.0
pillow==11.2.1
pluggy==1.6.0
pycparser==2.22
pydyf==0.11.0
pyphen==0.17.2
pytest==8.3.5
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.3
sqlalchemy==2.0.41
tinycss2==1.4.0
tinyhtml5==2.0.0
tomli==2.2.1
typing-extensions==4.13.2
urllib3==2.4.0
weasyprint==65.1
webencodings==0.5.1
werkzeug==3.1.3
zipp==3.21.0
zopfli==0.2.3.post1
psycopg2-binary==2.9.10
